Added a C++ program 'IncomeTaxCalculator' and did some file arrangement
