+++ README.md
@@ -87,3 +87,4 @@ Vatsla Adhiakri <br/>
 Sana Shah <br/>
 Archit Gupta <br/>
 Samrat Khan <br/>
+Cyrus Bhandari <br/>
